refactor(windows): consolidate unsafe boundary
